+++ files/gitlab-cookbooks/registry/attributes/default.rb
@@ -32,6 +32,7 @@ default['registry']['health_storagedriver_enabled'] = true
 default['registry']['storage_delete_enabled'] = nil
 default['registry']['storage'] = nil
 default['registry']['debug_addr'] = nil
+default['registry']['validation_enabled'] = false
 default['registry']['autoredirect'] = false
 
 ####